+++ sysdeps/riscv/rvf/s_lroundf.c
@@ -18,11 +18,12 @@
 
 #include <math.h>
 #include <libm-alias-float.h>
+#include <stdint.h>
 
 long int
 __lroundf (float x)
 {
-  long int res;
+  int32_t res;
   asm ("fcvt.w.s %0, %1, rmm" : "=r" (res) : "f" (x));
   return res;
 }
